Changed to a VCS Project and added the current files that builds but does nothing yet.
