+++ lab4/lab4.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.31101.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lab4", "lab4\lab4.vcxproj", "{9524AFB0-76F0-40A9-9985-7F7D4BB2850A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Win32 = Debug|Win32
+		Debug|x64 = Debug|x64
+		Release|Win32 = Release|Win32
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9524AFB0-76F0-40A9-9985-7F7D4BB2850A}.Debug|Win32.ActiveCfg = Debug|Win32
+		{9524AFB0-76F0-40A9-9985-7F7D4BB2850A}.Debug|Win32.Build.0 = Debug|Win32
+		{9524AFB0-76F0-40A9-9985-7F7D4BB2850A}.Debug|x64.ActiveCfg = Debug|x64
+		{9524AFB0-76F0-40A9-9985-7F7D4BB2850A}.Debug|x64.Build.0 = Debug|x64
+		{9524AFB0-76F0-40A9-9985-7F7D4BB2850A}.Release|Win32.ActiveCfg = Release|Win32
+		{9524AFB0-76F0-40A9-9985-7F7D4BB2850A}.Release|Win32.Build.0 = Release|Win32
+		{9524AFB0-76F0-40A9-9985-7F7D4BB2850A}.Release|x64.ActiveCfg = Release|x64
+		{9524AFB0-76F0-40A9-9985-7F7D4BB2850A}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
