--- lab4/main.cl
@@ -1,4 +1,0 @@
-void kernel pixelCalculation()
-{
-
-}
