--- lab4/main.h
@@ -1,17 +1,0 @@
-//
-// Created by secksx on 3/27/15.
-//
-
-#ifndef LAB4_MAIN_H
-#define LAB4_MAIN_H
-
-#include <CL/cl.hpp>
-#include <stdio.h>
-#include <fstream>
-
-class main {
-    int main(int argc, char* argv[]);
-};
-
-
-#endif //LAB4_MAIN_H
